% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conus_Albers_Check.R
\name{con_conus_albers}
\alias{con_conus_albers}
\title{con_conus_albers}
\usage{
con_conus_albers(spat_object)
}
\arguments{
\item{spat_object}{file of type spatial polygon, line, or points. Also accepts sf objects}
}
\description{
Check projection of a function and convert to BC conus albers
}
\note{
Author: Sarah Smith-Tripp, November 2021
}
