% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irss_gantt.R
\name{irss_gantt}
\alias{irss_gantt}
\title{Generate Gantt Chart}
\usage{
irss_gantt(
  gantt_df,
  palette = ggplot2::scale_colour_discrete(),
  transparent = FALSE,
  ...
)
}
\arguments{
\item{gantt_df}{A dataframe useable by ggplot; see \code{\link[iRss]{gantt_example}} for format}

\item{palette}{a color palette function to be used to colour the plot}

\item{transparent}{makes the background transparent, need to save it to properly visualize}

\item{...}{ggplot theme options}
}
\description{
Generate Gantt Chart
}
\examples{
# change palette
irss_gantt(gantt_df = gantt_example, palette = scico::scale_colour_scico_d(palette = "batlow"))

# make background transparent
irss_gantt(gantt_df = gantt_example, transparent = TRUE)


}
