% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_in_chunks.R
\name{mosaic_in_chunks}
\alias{mosaic_in_chunks}
\title{Mosaic small tiles to larger ones}
\usage{
mosaic_in_chunks(
  basedir,
  odir,
  file_type,
  idx = NULL,
  chunk_size = NULL,
  out_file_type = ".tif",
  pixel_type = "32_BIT_FLOAT",
  cellsize = "",
  number_of_bands = 1,
  mosaic_method = "MEAN",
  mosaic_colormap_mode = "FIRST"
)
}
\arguments{
\item{basedir}{Character String. Path to input files}

\item{odir}{Character String. Path to output directory (Will be created if it doesnt exist)}

\item{file_type}{Character String. File identifier (e.g. ".tif" / "avg.tif")}

\item{idx}{Character String. String to append to output mosaics}

\item{chunk_size}{Number of files per chunk (defaults to all)}

\item{out_file_type}{Character String. Output file type e.g. ".tif" / ".bil" see
https://desktop.arcgis.com/en/arcmap/10.3/tools/data-management-toolbox/mosaic-to-new-raster.htm}

\item{pixel_type}{default: \code{"32_BIT_FLOAT"} - see
\href{https://desktop.arcgis.com/en/arcmap/10.3/tools/data-management-toolbox/mosaic-to-new-raster.htm}{MosaictoNewRaster}}

\item{cellsize}{default: \code{""} - see
\href{https://desktop.arcgis.com/en/arcmap/10.3/tools/data-management-toolbox/mosaic-to-new-raster.htm}{MosaictoNewRaster}}

\item{number_of_bands}{default: \code{1} - see
\href{https://desktop.arcgis.com/en/arcmap/10.3/tools/data-management-toolbox/mosaic-to-new-raster.htm}{MosaictoNewRaster}}

\item{mosaic_method}{default: \code{"MEAN"} - see
\href{https://desktop.arcgis.com/en/arcmap/10.3/tools/data-management-toolbox/mosaic-to-new-raster.htm}{MosaictoNewRaster}}

\item{mosaic_colormap_mode}{default: \code{"FIRST"} - see
\href{https://desktop.arcgis.com/en/arcmap/10.3/tools/data-management-toolbox/mosaic-to-new-raster.htm}{MosaictoNewRaster}}
}
\description{
Mosaic raster tiles into chunks using arcmap MosaicToNewRaster and \link[reticulate]{import}
}
\note{
Mosaics files in \code{basedir} using mosaic to new raster functionality in arcmap
}
\examples{
\dontrun{

#--- define path of ARCGIS python on your PC ---#
arcpy3_dir <- 'C:/Program Files/ArcGIS/Pro/bin/Python'

#--- specify location of python for reticulate ---#
use_condaenv(condaenv = file.path(arcpy3_dir, 'envs', 'arcgispro-py3'),
             conda    = file.path(arcpy3_dir, 'Scripts' ,'conda.exe'),
             required = TRUE)

arcpy <- reticulate::import("arcpy")

mosaic_in_chunks(basedir = "F:/_2021_SSCALS/03_raster/dtm",
                 odir = "F:/_2021_SSCALS/03_raster/dtm/mosaics",
                 file_type = ".bil",
                 idx = "dtm",
                 chunk_size = 1000)


}

}
\author{
Tristan R.H. Goodbody
}
