% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_transparentdark.R
\name{theme_transparentdark}
\alias{theme_transparentdark}
\title{Make a transparent plot in ggplot}
\usage{
theme_transparentdark()
}
\description{
Make a transparent plot in ggplot
}
\examples{
# change palette
irss_gantt(gantt_df = gantt_example, palette = scico::scale_colour_scico_d(palette = "batlow")) +
theme_transparentdark()

NOTE: if you would like the labels to be white you must specify this in the geom_label of the plot argument
e.g. : geom_label(aes(), color = "white")

# make background transparent
irss_gantt(gantt_df = gantt_example) + theme_transparentdark()


}
