% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irss_gantt.R
\name{irss_gantt}
\alias{irss_gantt}
\title{Generate Gantt Chart}
\usage{
irss_gantt(gantt_df, palette = ggplot2::scale_colour_discrete(), ...)
}
\arguments{
\item{gantt_df}{A dataframe useable by ggplot; see \code{\link[iRss]{gantt_example}} for format}

\item{palette}{a color palette function to be used to colour the plot}

\item{...}{ggplot theme options and geom line options, used to make transparent background or other minor edits}
}
\description{
Generate Gantt Chart
}
